refactor: share Claude conversion helpers between anthropic and vertexai
